rapid_law/components/graph_components/state.py
from typing import Annotated
from langgraph.graph.message import add_messages
from typing_extensions import TypedDict


class State(TypedDict):
    messages: Annotated[list, add_messages]
    docs: list[dict]
    prompt: str
