pyproject.toml
[tool.poetry]
name = "rapid-law"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.29.5"
langgraph = "^0.1.19"
pinecone-client = {extras = ["grpc"], version = "^5.0.1"}
streamlit = "^1.37.0"
python-dotenv = "^1.0.1"
cohere = "^5.6.2"
langchain-text-splitters = "^0.2.2"
tiktoken = "^0.7.0"
google-generativeai = "^0.7.2"
ollama = "^0.3.1"
autoreload = "^0.1.2"
hugchat = "^0.4.10"
elementpath = "^4.4.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
